Move arcface_loss into loss_utils.py.
